requirements.txt
streamlit==1.28.1
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.3.7
faiss-cpu==1.7.4
unstructured==0.11.8
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
numpy==1.24.3
pypdf==3.17.1
beautifulsoup4==4.12.2
lxml==4.9.3
tiktoken==0.5.1
transformers==4.41.2
sentence-transformers==2.7.0
huggingface-hub==0.23.2 